+++ src/NextUnit.Generator/Helpers/DataSourceAttributeReader.cs
@@ -92,17 +92,8 @@ internal static class DataSourceAttributeReader
 
         foreach (var attribute in methodSymbol.GetAttributes())
         {
-            var attrClass = attribute.AttributeClass;
-            if (attrClass is not { IsGenericType: true })
-            {
-                continue;
-            }
-
-            var constructedFrom = attrClass.ConstructedFrom;
-            var metadataName = constructedFrom.MetadataName;
-
-            if (!metadataName.StartsWith(NextUnitAttributeNames.MetadataNames.ClassDataSourceAttributePrefix, StringComparison.Ordinal) ||
-                constructedFrom.ContainingNamespace.ToDisplayString() != NextUnitAttributeNames.Namespace)
+            var sourceTypes = ClassDataSourceAttributeMatcher.GetClassDataSourceTypes(attribute);
+            if (sourceTypes.IsEmpty)
             {
                 continue;
             }
@@ -122,7 +113,7 @@ internal static class DataSourceAttributeReader
                 }
             }
 
-            foreach (var typeArg in attrClass.TypeArguments)
+            foreach (var typeArg in sourceTypes)
             {
                 var typeName = typeArg.ToDisplayString(AttributeHelper.TypeExpressionFormat);
                 builder.Add(new ClassDataSource(typeName, sharedType, key));
@@ -267,44 +258,35 @@ internal static class DataSourceAttributeReader
                     unreachableMemberTypeName: unreachableMemberTypeName);
             }
 
-            var attrClass = attribute.AttributeClass;
-            if (attrClass is { IsGenericType: true })
+            if (ClassDataSourceAttributeMatcher.GetValuesFromType(attribute) is { } typeArg)
             {
-                var constructedFrom = attrClass.ConstructedFrom;
-                var metadataName = constructedFrom.MetadataName;
+                var classTypeName = typeArg.ToDisplayString(AttributeHelper.TypeExpressionFormat);
+                var sharedType = 0;
+                var key = (string?)null;
 
-                if (metadataName.StartsWith(NextUnitAttributeNames.MetadataNames.ValuesFromAttributePrefix, StringComparison.Ordinal) &&
-                    constructedFrom.ContainingNamespace.ToDisplayString() == NextUnitAttributeNames.Namespace)
+                foreach (var namedArg in attribute.NamedArguments)
                 {
-                    var typeArg = attrClass.TypeArguments[0];
-                    var classTypeName = typeArg.ToDisplayString(AttributeHelper.TypeExpressionFormat);
-                    var sharedType = 0;
-                    var key = (string?)null;
-
-                    foreach (var namedArg in attribute.NamedArguments)
+                    if (namedArg.Key == "Shared" && namedArg.Value.Value is int sharedValue)
                     {
-                        if (namedArg.Key == "Shared" && namedArg.Value.Value is int sharedValue)
-                        {
-                            sharedType = sharedValue;
-                        }
-                        else if (namedArg.Key == "Key" && namedArg.Value.Value is string keyValue)
-                        {
-                            key = keyValue;
-                        }
+                        sharedType = sharedValue;
                     }
-
-                    return new ParameterDataSourceDescriptor(
-                        parameterIndex: index,
-                        parameterName: parameter.Name,
-                        kind: ParameterDataSourceKind.Class,
-                        inlineValues: EquatableArray<ConstantValue>.Empty,
-                        memberName: null,
-                        memberTypeName: null,
-                        memberKind: DataSourceMemberKind.Unknown,
-                        classTypeName: classTypeName,
-                        sharedType: sharedType,
-                        sharedKey: key);
+                    else if (namedArg.Key == "Key" && namedArg.Value.Value is string keyValue)
+                    {
+                        key = keyValue;
+                    }
                 }
+
+                return new ParameterDataSourceDescriptor(
+                    parameterIndex: index,
+                    parameterName: parameter.Name,
+                    kind: ParameterDataSourceKind.Class,
+                    inlineValues: EquatableArray<ConstantValue>.Empty,
+                    memberName: null,
+                    memberTypeName: null,
+                    memberKind: DataSourceMemberKind.Unknown,
+                    classTypeName: classTypeName,
+                    sharedType: sharedType,
+                    sharedKey: key);
             }
         }